Helper functions moved to an external utility module.
